requirements.txt
# Django and Core Packages
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0

# MongoDB Integration
mongoengine==0.28.2
pymongo==3.12.3

# REMOVED: Google OAuth Authentication packages
# django-allauth==0.63.6
# google-auth==2.32.0
# google-auth-oauthlib==1.2.1
# google-auth-httplib2==0.2.0
# requests-oauthlib==2.0.0

# Environment and Configuration
python-dotenv==1.0.1
python-decouple==3.8

# Image Processing
Pillow==10.4.0

# Security and Authentication (for admin password management)
cryptography==41.0.7

# Machine Learning for Random Forest Algorithm
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4

